+++ CHANGELOG.md
@@ -69,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Limit WebSocket reconnect attempts
 - Refactored default screen sharing session start to fix state corruption bug
 - Update Travis script to separate unit and integration tests into different jobs.
+- Validate session Id and disable extendedDebugging flag for SauceLabs
 
 ### Removed
 - Remove SDP class withPlanBSimulcast method